Finished Sessions for the most part and begun on some formatting stuff: Can push (#25)
* Finished up on sessions and made the changed which were suggested.
Did some small formatting stuff on the home buttons
Reformatted the favorites sections and favorites star button, so I can begin with working or the favorites section folders.
added the cordinatesto the users table.

(This is unfinished): I plan on working on fixing the history as well as getting a start on the favorites sections tommorrow)

* Fixed History.. Fixed sessions making too many calls

* Hopefully fixed too many api calls

* small history fixes: more to be changed